@@ -4,7 +4,7 @@ publish_to: "none"
 version: 0.1.0
 
 environment:
-  sdk: ^3.5.4
+  sdk: ^3.7.0
 
 dependencies:
   flutter:
@@ -18,6 +18,11 @@ dependencies:
   firebase_auth: ^5.5.1
   cloud_firestore: ^5.6.5
 
+  # Firebase Social Auth
+  google_sign_in: ^6.3.0
+  sign_in_with_apple: ^6.1.4
+  flutter_facebook_auth: ^7.1.1
+
   # Google Maps
   google_maps_flutter: ^2.10.1
   geocoding: ^3.0.0
@@ -29,6 +34,8 @@ dependencies:
   # Utilities
   intl: ^0.20.2
   logger: ^2.5.0
+  crypto: ^3.0.6
+  flutter_dotenv: ^5.2.1
 
   # UI Components
   calendar_date_picker2: ^1.1.9
